+++ Rebus/Sagas/CorrelationProperty.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace Rebus.Sagas
@@ -28,27 +29,39 @@ namespace Rebus.Sagas
         /// <param name="sagaType">Specifies the saga type (i.e. the handler type) that contains the logic of the saga</param>
         public CorrelationProperty(Type messageType, Func<object, object> valueFromMessage, Type sagaDataType, string propertyName, Type sagaType)
         {
-            if (messageType == null) throw new ArgumentNullException("messageType");
-            if (sagaDataType == null) throw new ArgumentNullException("sagaDataType");
-            if (propertyName == null) throw new ArgumentNullException("propertyName");
-            if (sagaType == null) throw new ArgumentNullException("sagaType");
+            if (messageType == null) throw new ArgumentNullException(nameof(messageType));
+            if (sagaDataType == null) throw new ArgumentNullException(nameof(sagaDataType));
+            if (propertyName == null) throw new ArgumentNullException(nameof(propertyName));
+            if (sagaType == null) throw new ArgumentNullException(nameof(sagaType));
             PropertyName = propertyName;
             SagaType = sagaType;
             ValueFromMessage = valueFromMessage;
             SagaDataType = sagaDataType;
             MessageType = messageType;
-
             Validate();
         }
 
         void Validate()
         {
-            var propertyType = SagaDataType.GetProperty(PropertyName).PropertyType;
+            if (string.IsNullOrWhiteSpace(PropertyName))
+            {
+                throw new ArgumentException($"Reflected saga data correlation property name from {SagaDataType} is empty! This is most likely because the expression passed to the correlation configuration could not be properly reflected - it's the part indicated by !!! here: config.Correlate<TMessage>(m => m.SomeField, d => !!!) - please be sure that you are pointing to a simple property of the saga data");
+            }
+
+            var sagaDataProperty = SagaDataType.GetProperty(PropertyName);
+
+            if (sagaDataProperty == null)
+            {
+                throw new ArgumentException($"Could not find correlation property '{PropertyName}' on saga data of type {SagaDataType}!");
+            }
+
+            var propertyType = sagaDataProperty.PropertyType;
 
             if (AllowedCorrelationPropertyTypes.Contains(propertyType)) return;
 
-            throw new ArgumentException(string.Format("Cannot correlate with the '{0}' property on the '{1}' saga data type - only allowed types are: {2}",
-                PropertyName, SagaDataType.Name, string.Join(", ", AllowedCorrelationPropertyTypes.Select(t => t.Name))));
+            var allowedTypes = string.Join(", ", AllowedCorrelationPropertyTypes.Select(t => t.Name));
+
+            throw new ArgumentException($"Cannot correlate with the '{PropertyName}' property on the '{SagaDataType.Name}' saga data type - only allowed types are: {allowedTypes}");
         }
 
         /// <summary>
@@ -64,16 +77,16 @@ namespace Rebus.Sagas
         /// <summary>
         /// Gets the type of the saga's saga data
         /// </summary>
-        public Type SagaDataType { get; private set; }
+        public Type SagaDataType { get; }
 
         /// <summary>
         /// Gets the name of the correlation property
         /// </summary>
-        public string PropertyName { get; private set; }
+        public string PropertyName { get; }
         
         /// <summary>
         /// The saga type (i.e. the handler type) that contains the logic of the saga
         /// </summary>
-        public Type SagaType { get; private set; }
+        public Type SagaType { get; }
     }
 }